Fix char encoding (sign extension issue)

+++ src/mprEncode.c
@@ -102,7 +102,7 @@ char *mprUriEncode(MprCtx ctx, cchar *inbuf, int map)
     mprAssert(inbuf);
 
     for (len = 1, ip = inbuf; *ip; ip++, len++) {
-        if (charMatch[(int) *ip] & map) {
+        if (charMatch[(int) (uchar) *ip] & map) {
             len += 2;
         }
     }
@@ -186,7 +186,7 @@ char *mprEscapeCmd(MprCtx ctx, cchar *cmd, int escChar)
     mprAssert(cmd);
 
     for (len = 1, ip = cmd; *ip; ip++, len++) {
-        if (charMatch[(int) *ip] & MPR_ENCODE_SHELL) {
+        if (charMatch[(int) (uchar) *ip] & MPR_ENCODE_SHELL) {
             len++;
         }
     }
@@ -227,7 +227,7 @@ char *mprEscapeHtml(MprCtx ctx, cchar *html)
     int     len;
 
     for (len = 1, ip = html; *ip; ip++, len++) {
-        if (charMatch[(int) *ip] & MPR_ENCODE_HTML) {
+        if (charMatch[(int) (uchar) *ip] & MPR_ENCODE_HTML) {
             len += 5;
         }
     }
